+++ src/common/redis.ts
@@ -19,9 +19,10 @@ export const redisSubscriber = new Redis(config.redisUrl, {
 });
 
 export const rateLimitRedis = new Redis(config.rateLimitRedisUrl, {
-  maxRetriesPerRequest: null,
+  maxRetriesPerRequest: 1,
   enableReadyCheck: false,
   enableOfflineQueue: false,
+  commandTimeout: 3000,
 });
 
 // https://redis.io/topics/distlock