+++ src/jobs/order-updates/by-maker-queue.ts
@@ -64,7 +64,7 @@ if (config.doBackgroundWork) {
                     ELSE 'no-balance'
                   END)::order_fillability_status_t AS new_status,
                   (CASE
-                    WHEN ft_balances.amount >= (orders.price * orders.quantity_remaining) THEN upper(orders.valid_between)
+                    WHEN ft_balances.amount >= (orders.price * orders.quantity_remaining) THEN nullif(upper(orders.valid_between), 'infinity')
                     ELSE to_timestamp($/timestamp/)
                   END)::timestamptz AS expiration
                 FROM orders
@@ -266,7 +266,7 @@ if (config.doBackgroundWork) {
                     ELSE 'no-balance'
                   END)::order_fillability_status_t AS new_status,
                   (CASE
-                    WHEN nft_balances.amount >= orders.quantity_remaining THEN upper(orders.valid_between)
+                    WHEN nft_balances.amount >= orders.quantity_remaining THEN nullif(upper(orders.valid_between), 'infinity')
                     ELSE to_timestamp($/timestamp/)
                   END)::TIMESTAMPTZ AS expiration
                 FROM orders
@@ -293,8 +293,8 @@ if (config.doBackgroundWork) {
             // Filter any orders that didn't change status
             const values = fillabilityStatuses
               .filter(({ old_status, new_status }) => old_status !== new_status)
-              // Exclude 'foundation' orders (which need special rules because of the escrowed orderbook)
-              .filter(({ kind }) => kind !== "foundation")
+              // Exclude escrowed orders
+              .filter(({ kind }) => kind !== "foundation" && kind !== "cryptopunks")
               // When a token gets transferred, X2Y2 will off-chain cancel all the
               // orders from the initial owner, so that if they ever get the token
               // back in their wallet no order will get reactivated (they are able
@@ -365,7 +365,7 @@ if (config.doBackgroundWork) {
                       ELSE 'no-approval'
                     END)::order_approval_status_t AS new_status,
                     (CASE
-                      WHEN nft_approval_events.approved THEN upper(orders.valid_between)
+                      WHEN nft_approval_events.approved THEN nullif(upper(orders.valid_between), 'infinity')
                       ELSE to_timestamp($/timestamp/)
                     END)::TIMESTAMPTZ AS expiration
                   FROM nft_approval_events
@@ -392,8 +392,8 @@ if (config.doBackgroundWork) {
             // Filter any orders that didn't change status
             const values = approvalStatuses
               .filter(({ old_status, new_status }) => old_status !== new_status)
-              // Exclude 'foundation' orders (which need special rules because of the escrowed orderbook)
-              .filter(({ kind }) => kind !== "foundation")
+              // Exclude escrowed orders
+              .filter(({ kind }) => kind !== "foundation" && kind !== "cryptopunks")
               .map(({ id, new_status, expiration }) => ({
                 id,
                 approval_status: new_status,